Extracting common issue monitoring logic into its own class Inlining the functional interface as it wont be used anywhere else.
