--- src/main/java/io/spring/issuebot/IssueOperation.java
@@ -1,10 +1,0 @@
-package io.spring.issuebot;
-
-import io.spring.issuebot.github.Issue;
-
-@FunctionalInterface
-public interface IssueOperation {
-
-	void run(Issue issue, IssueListener issueListener);
-
-}
