+++ src/main/java/io/spring/issuebot/github/ClosedIssuesListener.java
@@ -1,9 +1,8 @@
 package io.spring.issuebot.github;
 
-import io.spring.issuebot.IssueListener;
+import io.spring.issuebot.IssueMonitor;
 import io.spring.issuebot.Repository;
 import io.spring.issuebot.RepositoryListener;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -13,13 +12,13 @@ public class ClosedIssuesListener implements RepositoryListener {
 
 	private static final Logger log = LoggerFactory.getLogger(ClosedIssuesListener.class);
 
-	private final List<IssueListener> issueListeners;
-
 	private final GitHubOperations gitHub;
 
+	private final IssueMonitor issueMonitor;
+
 	public ClosedIssuesListener(GitHubOperations gitHub,
-			List<IssueListener> issueListeners) {
-		this.issueListeners = issueListeners;
+			IssueMonitor issueMonitor) {
+		this.issueMonitor = issueMonitor;
 		this.gitHub = gitHub;
 	}
 
@@ -29,22 +28,8 @@ public class ClosedIssuesListener implements RepositoryListener {
 			Page<Issue> page = this.gitHub.getClosedIssuesWithLabel(
 					repository.getOrganization(), repository.getName(),
 					"status: waiting-for-triage");
-			while (page != null) {
-				for (Issue issue : page.getContent()) {
-					for (IssueListener issueListener : this.issueListeners) {
-						try {
-							issueListener.onIssueClosure(repository, issue);
-						}
-						catch (Exception ex) {
-							log.warn("Listener '{}' failed when handling issue '{}'",
-									issueListener, issue, ex);
-						}
-					}
-				}
-				page = page.next();
-			}
-		}
-		catch (Exception ex) {
+			issueMonitor.monitorIssue(page, (issue, issueListener) -> issueListener.onIssueClosure(repository, issue));
+		} catch (Exception ex) {
 			log.warn("A failure occurred during monitoring of {}/{}",
 					repository.getOrganization(), repository.getName(), ex);
 		}
