+++ src/main/java/io/spring/issuebot/github/OpenIssuesListener.java
@@ -1,9 +1,8 @@
 package io.spring.issuebot.github;
 
-import io.spring.issuebot.IssueListener;
+import io.spring.issuebot.IssueMonitor;
 import io.spring.issuebot.Repository;
 import io.spring.issuebot.RepositoryListener;
-import java.util.List;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
@@ -13,13 +12,12 @@ public class OpenIssuesListener implements RepositoryListener {
 
 	private static final Logger log = LoggerFactory.getLogger(OpenIssuesListener.class);
 
-	private final List<IssueListener> issueListeners;
-
 	private final GitHubOperations gitHub;
+	private final IssueMonitor issueMonitor;
 
 	public OpenIssuesListener(GitHubOperations gitHub,
-			List<IssueListener> issueListeners) {
-		this.issueListeners = issueListeners;
+			IssueMonitor issueMonitor) {
+		this.issueMonitor = issueMonitor;
 		this.gitHub = gitHub;
 	}
 
@@ -28,20 +26,7 @@ public class OpenIssuesListener implements RepositoryListener {
 		try {
 			Page<Issue> page = this.gitHub.getIssues(repository.getOrganization(),
 					repository.getName());
-			while (page != null) {
-				for (Issue issue : page.getContent()) {
-					for (IssueListener issueListener : this.issueListeners) {
-						try {
-							issueListener.onOpenIssue(repository, issue);
-						}
-						catch (Exception ex) {
-							log.warn("Listener '{}' failed when handling issue '{}'",
-									issueListener, issue, ex);
-						}
-					}
-				}
-				page = page.next();
-			}
+			 issueMonitor.monitorIssue(page, (issue, issueListener) -> issueListener.onOpenIssue(repository, issue));
 		}
 		catch (Exception ex) {
 			log.warn("A failure occurred during monitoring of {}/{}",
